+++ src/module.ts
@@ -1,2 +1,3 @@
+export * from './ansi.js';
 export * from './logger.js';
 export * from './stringify.js';